KAS-1402: refactor refreshData method
