+++ app/components/utils/add-minister-modal/component.js
@@ -1,33 +1,12 @@
 import Component from '@ember/component';
 import EmberObject from '@ember/object';
-
+import {
+  refreshData
+} from '../../../utils/manage-minister-util';
 export default Component.extend({
 
   rowToShow: null,
   selectedMandatee: null,
-
-  async refreshData(mandatee) {
-    this.set('isLoading', true);
-
-    const iseCodes = (await mandatee.get('iseCodes')).filter((item) => item);
-    const fields = (await Promise.all(iseCodes.map((iseCode) => iseCode.get('field')))).filter((item) => item);
-    const domains = await Promise.all(fields.map((field) => field.get('domain')));
-
-    const rowToShow = EmberObject.create({
-      domains: [...new Set(domains)],
-      fields: [...new Set(fields)]
-    });
-    rowToShow.get('domains').map((domain) => domain.set('selected', false));
-    rowToShow.get('fields').map((domain) => domain.set('selected', false));
-    const mandateeRows = await this.get('mandateeRows');
-
-    if (!mandateeRows) {
-      rowToShow.set('isSubmitter', true);
-    }
-    this.set('isLoading', false);
-    return rowToShow;
-  },
-
 
   actions: {
 
@@ -81,7 +60,10 @@ export default Component.extend({
 
     async mandateeSelected(mandatee) {
       this.set('selectedMandatee', mandatee);
-      this.set('rowToShow', await this.refreshData(mandatee));
+      this.set('isLoading', true);
+      const rowsToShow = await refreshData(mandatee,await this.get('mandateeRows'));
+      this.set('rowToShow',rowsToShow);
+      this.set('isLoading', false);
     },
 
     async selectDomain(domain, value) {
